requirements.txt
numpy==2.0.1
opencv_contrib_python==4.6.0.66
paddlepaddle
paddleocr==2.7.0.3
paddleocr.egg==info
pyserial==3.5
python-socketio==5.11.3
setuptools==69.5.1
setuptools==68.1.2
